Animated time separators
Now, the : in HH:MM:SS is animated, it goes up and down.

As a bonus, the code has been slightly refactored into functions. Not
perfect, but good enough.

In special, the display_interfaces needs the delay inside itself, while
the other functions return immediately and have the delay outside.
That's inconsistent, but I haven't thought of a better solution yet.