Display time difference from some reference time
There are two algorithms: count time based on how many seconds, or
count time based on the calendar months ans years. One is more precise,
the other is our traditional "age".

The reference time can be in the past or in the future. It can even be
in another timezone.

(This commit was mostly written during awake hours in the night, using a
tablet running Termux to ssh to a Raspberry Pi.)