Removing netifaces dependency
I can get the same results using `psutil`, which I added as a dependency
in the previous commit. Also, `netifaces` seems abandoned/archived.
Performance-wise, both modules seem to be implemented in C.

This also gets rid of a racing condition that could crash this script.
If a network interface becomes unavailable between fetching the list of
interfaces and querying their addresses, the code would throw an
exception and abort. Now, everything is fetched at once, and the racing
condition is taken care inside the `psutil` library. (Or, if not taken
care, at least it's their responsibility, not mine anymore.)

I've also filtered out the docker interfaces, as those are mostly
useless.